+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 ## 1.7 branch
 ### 1.7.4
 * all methods provided by the `Filesystem` class (except for `makePathRelative()`) can now be called statically;
+* added `rtr()` global function, a fast and convenient alias for `Filesystem::rtr()`;
 * `objects_map()` global function is now deprecated and will be removed in a later release.
 
 ### 1.7.3